+++ lib/screen/post_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import './main.dart';
+import '../main.dart';
 
 class PostPage extends StatefulWidget {
   @override
